resume - Copy/resume - Copy/requirements .txt

flask
pdfplumber
python-docx
reportlab
nltk
torch
numpy
scikit-learn
pandas
